requirements.txt
flask
flask-socketio
python-socketio
flask-login
flask-sqlalchemy
flask-migrate
jinja2
werkzeug
python-dotenv
pytz

# .\.venv\Scripts\pip install -r requirements.txt